# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tenders API Documentation",
        "description": "Tenders API Documentation",
        "contact": {
            "email": "info@just.sd"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Tenders API Server"
        }
    ],
    "paths": {
        "/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get list of clients",
                "description": "get all clients",
                "operationId": "get_all_clients",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "update client",
                "description": "update client",
                "operationId": "update_client",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "client_name",
                        "in": "path",
                        "description": "Client Name",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "phone",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "city",
                        "in": "path",
                        "description": "city",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "state",
                        "in": "path",
                        "description": "State",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "country",
                        "in": "path",
                        "description": "country",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Address",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "store client",
                "description": "store client",
                "operationId": "store_client",
                "parameters": [
                    {
                        "name": "client_name",
                        "in": "path",
                        "description": "Client Name",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "phone",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "city",
                        "in": "path",
                        "description": "city",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "state",
                        "in": "path",
                        "description": "State",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "country",
                        "in": "path",
                        "description": "country",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Address",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "show client",
                "description": "show client",
                "operationId": "show_client",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "delete client",
                "description": "delete client",
                "operationId": "delete_client",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}